Issue 114 - SampleInstrument sometimes produces crackling sound (#116)
* Optimized sequencers event lookup
* Fixed long standing issue with overlapping events that occurred when a sequence would loop
* Added unit tests for changes
* AudioEngineProps that must be unsigned are now declared as such
* Activated deprecation warnings for BaseAudioEvent, removed deprecated functions from SWIG wrapper
* Fixed issue with deadlocks on instrument event vector locks
* Fixed issue where tempo adjustments led to undefined behaviour for event offets
* Added x86_64 specific compiler optimizations
* Introduced sustained performance mode (from Android N onwards)
* Added load stabilizer to prevent CPU frequency scaling
* Set thread affinity to prevent CPU core migration
* Thread affinity is assigned to exclusive cores (from Android N onwards)
* Use direct index operator on stack vectors, pass by reference in function arguments
* Fix playback issue that would occur when samples playback rate would be below 1 (e.g. pitch shifted down)

+++ src/main/cpp/processingchain.cpp
@@ -47,7 +47,7 @@ void ProcessingChain::addProcessor( BaseProcessor* aProcessor )
 
 void ProcessingChain::removeProcessor( BaseProcessor* aProcessor )
 {
-    for ( int i = 0; i < _activeProcessors.size(); i++ )
+    for ( size_t i = 0; i < _activeProcessors.size(); i++ )
     {
         if ( _activeProcessors.at( i ) == aProcessor )
         {
